move null logger to a new file
