+++ src/common/logging/index.ts
@@ -1,6 +1,7 @@
 export { LogId } from "./loggingDefinitions.js";
 export type * from "./loggingTypes.js";
 export * from "./loggerBase.js";
+export { NullLogger } from "./nullLogger.js";
 export { CompositeLogger } from "./compositeLogger.js";
 export { ConsoleLogger } from "./consoleLogger.js";
 export { DiskLogger } from "./diskLogger.js";
