+++ src/common/logging/nullLogger.ts
@@ -1,0 +1,14 @@
+import type { LoggerType } from "./loggingTypes.js";
+import { LoggerBase } from "./loggerBase.js";
+
+export class NullLogger extends LoggerBase {
+    protected type?: LoggerType;
+
+    constructor() {
+        super(undefined);
+    }
+
+    protected logCore(): void {
+        // No-op logger, does not log anything
+    }
+}
